requirement.txt
pydantic
fastapi
uvicorn
python-dotenv